refactor: props로 필수 값만 받게 수정
props로 필수 값만 받게 수정
